+++ tests/integration/devtools/test_hot_reload.py
@@ -1,6 +1,7 @@
 import os
 from time import sleep
 
+from dash.testing.wait import until
 import dash_html_components as html
 import dash
 from dash.dependencies import Input, Output
@@ -12,6 +13,24 @@ RED_BG = """
     background-color: red;
 }
 """
+
+GOUDA = """
+window.cheese = 'gouda';
+"""
+
+
+def replace_file(filename, new_content):
+    path = os.path.join(
+        os.path.dirname(__file__), "hr_assets", filename
+    )
+    with open(path, "r+") as fp:
+        sleep(1)  # ensure a new mod time
+        old_content = fp.read()
+        fp.truncate(0)
+        fp.seek(0)
+        fp.write(new_content)
+
+    return path, old_content
 
 
 def test_dvhr001_hot_reload(dash_duo):
@@ -42,15 +61,12 @@ def test_dvhr001_hot_reload(dash_duo):
         "#hot-reload-content", "background-color", "rgba(0, 0, 255, 1)"
     )
 
-    hot_reload_file = os.path.join(
-        os.path.dirname(__file__), "hr_assets", "hot_reload.css"
-    )
-    with open(hot_reload_file, "r+") as fp:
-        sleep(1)  # ensure a new mod time
-        old_content = fp.read()
-        fp.truncate(0)
-        fp.seek(0)
-        fp.write(RED_BG)
+    # set a global var - if we soft reload it should still be there,
+    # hard reload will delete it
+    dash_duo.driver.execute_script("window.someVar = 42;")
+    assert dash_duo.driver.execute_script("return window.someVar") == 42
+
+    soft_reload_file, old_soft = replace_file("hot_reload.css", RED_BG)
 
     try:
         # red is live changed during the test execution
@@ -59,12 +75,37 @@ def test_dvhr001_hot_reload(dash_duo):
         )
     finally:
         sleep(1)  # ensure a new mod time
-        with open(hot_reload_file, "w") as f:
-            f.write(old_content)
+        with open(soft_reload_file, "w") as f:
+            f.write(old_soft)
 
     dash_duo.wait_for_style_to_equal(
         "#hot-reload-content", "background-color", "rgba(0, 0, 255, 1)"
     )
+
+    # only soft reload, someVar is still there
+    assert dash_duo.driver.execute_script("return window.someVar") == 42
+
+    assert dash_duo.driver.execute_script("return window.cheese") == "roquefort"
+
+    hard_reload_file, old_hard = replace_file("hot_reload.js", GOUDA)
+
+    try:
+        until(
+            lambda: dash_duo.driver.execute_script("return window.cheese") == "gouda",
+            timeout=3
+        )
+    finally:
+        sleep(1)  # ensure a new mod time
+        with open(hard_reload_file, "w") as f:
+            f.write(old_hard)
+
+    until(
+        lambda: dash_duo.driver.execute_script("return window.cheese") == "roquefort",
+        timeout=3
+    )
+
+    # we've done a hard reload so someVar is gone
+    assert dash_duo.driver.execute_script("return window.someVar") is None
 
     # Now check the server status indicator functionality